refactor: server structure
